move `server`  resource concurrency handler to other file
